Move token to a specific step
