+++ src/test/groovy/GithubCheckStepTests.groovy
@@ -25,14 +25,11 @@ import static org.junit.Assert.assertTrue
 class GithubCheckStepTests extends ApmBasePipelineTest {
   def script
 
-  def privateKey
-
   @Override
   @Before
   void setUp() throws Exception {
     super.setUp()
     script = loadScript('vars/githubCheck.groovy')
-    privateKey = new File("src/test/resources/github-app-private-key-tests.pem").text
   }
 
   @Test
@@ -44,84 +41,6 @@ class GithubCheckStepTests extends ApmBasePipelineTest {
     }
     printCallStack()
     assertTrue(assertMethodCallContainsPattern('error', 'missing name argument'))
-    assertJobStatusFailure()
-  }
-
-  @Test
-  void test_with_wrong_vault_details() throws Exception {
-    try {
-      script.call(name: 'foo', secret: VaultSecret.SECRET_ERROR.toString())
-    } catch(err) {
-      // NOOP
-    }
-    printCallStack()
-    assertTrue(assertMethodCallContainsPattern('error', 'Unable to get credentials from the vault'))
-    assertJobStatusFailure()
-  }
-
-  @Test
-  void test_getToken_success() throws Exception {
-    helper.registerAllowedMethod('githubApiCall', [Map.class], {
-      return [token: 1]
-    })
-    script.getToken(jsonWebToken: 'foo', installationId: '123')
-    printCallStack()
-    assertTrue(assertMethodCallContainsPattern('githubApiCall', 'token=foo'))
-    assertTrue(assertMethodCallContainsPattern('githubApiCall', 'https://api.github.com/app/installations/123/access_tokens'))
-    assertJobStatusSuccess()
-  }
-
-  @Test
-  void test_getToken_failure() throws Exception {
-    helper.registerAllowedMethod('githubApiCall', [Map.class], {
-      throw new Exception('Forced a failure')
-    })
-    try {
-      script.getToken(jsonWebToken: 'foo', installationId: '123')
-    } catch(err) {
-      // NOOP
-    }
-    printCallStack()
-    assertTrue(assertMethodCallContainsPattern('error', 'getToken: Failed to create a JWT'))
-    assertJobStatusFailure()
-  }
-
-  @Test
-  void test_getRSAPrivateKey_success() throws Exception {
-    script.getRSAPrivateKey(privateKey)
-    printCallStack()
-    assertJobStatusSuccess()
-  }
-
-  @Test
-  void test_getRSAPrivateKey_failure() throws Exception {
-    try {
-      script.getRSAPrivateKey('secret')
-    } catch(err) {
-      // NOOP
-      println err
-    }
-    printCallStack()
-    assertTrue(assertMethodCallContainsPattern('error', 'getRSAPrivateKey: Failed to create a JWT'))
-    assertJobStatusFailure()
-  }
-
-  @Test
-  void test_getJsonWebToken_success() throws Exception {
-    script.getJsonWebToken(privateKeyContent: privateKey, appId: '123')
-    printCallStack()
-    assertJobStatusSuccess()
-  }
-
-  @Test
-  void test_getJsonWebToken_failure() throws Exception {
-    try {
-      script.getJsonWebToken(privateKeyContent: 'foo', appId: '123')
-    } catch(err) {
-      // NOOP
-    }
-    printCallStack()
-    assertTrue(assertMethodCallContainsPattern('error', 'getJsonWebToken: Failed to create a JWT'))
     assertJobStatusFailure()
   }
