+++ vars/githubAppToken.txt
@@ -1,0 +1,9 @@
+Get the GitHub APP token given the vault secret
+
+```
+def token = githubAppToken()
+```
+
+* secret: vault secret used to interact with the GitHub App, it should have the `key`, `installation_id` and `app_id` fields. Default: 'secret/observability-team/ci/github-app'
+
+[GitHub Check docs](https://docs.github.com/en/free-pro-team@latest/rest/reference/checks#runs)
